requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.29
PyJWT==2.8.0 